Cache Vendor public key for ten minutes
Caching for ten minutes may prevent some issues with repeated queries and speed up multiple requests.

It also will not be a problem (for long, at least!) if the Vendor cycles their keys.

Perhaps fetching the public key should happen as a background task so that encryption is performed faster. For a later time.…

Resolves #39