No longer store access keys in options DB
This was, I believe, a decision based on concerns for the time it takes to generate the hash. On 7.x it takes `.006` seconds. I'll perform additional timing on 5.x but since this is loaded once on one screen, we should be good to remove altogether.

Also, OptionKeys is no longer needed (and was never used…)