When no license key was defined, it was hashing an empty string
The SiteAccess::get_license_key() method is used as one part of a hashed value.

When no license key was set, the method would hash an empty string. This resulted in persistent values for the hash as well as the same hash being used for sites on the same network.

This is *not a security risk*, since even if the access key is the same, the support user only has access to the site they were created on.