If a user already exists, don't fail, use it
The user shouldn't already exist, but may if a deletion cron failed. It doesn't mean that the account creation process should fail.

What we don't want is for an attack vector to exist where a vendor declares their email as `very-secure@example.com` and then being able to modify or access that account in any way due to a simple match. That's why the match only is allowed when the hashing process has taken place.