Rename the log file structure
Added the client namespace to the log file. This will allow for log rotation in the future without affecting other TL namespaces.

Removed a few names from the file: The logs are already in a directory named `trustedlogin-logs`. we don't currently use logs per-level, so `-debug` doesn't make sense. We may store `vendor` logs here in the future, so prefixing with client remains a good idea.